Simplifications. Moved items rendering to the QueryBuilder component. Updated tests.
